move keytable > table pacakge and split into files
The splitting into two files, `keytable` and `table` was largely
to help me remember how this all worked, so when I introduce a new
table, I can simply copy `keyable` and reuse all the unexported
functions in `table`.
